refactor(cli): move plan exit logic to kilo tool
