+++ packages/opencode/src/kilocode/tool/plan.ts
@@ -1,0 +1,41 @@
+import { Effect, Schema } from "effect"
+import * as Tool from "@/tool/tool"
+import { InstanceState } from "@/effect/instance-state"
+import { Session } from "@/session/session"
+import { PlanFile } from "@/kilocode/plan-file"
+import EXIT_DESCRIPTION from "@/tool/plan-exit.txt"
+
+export const Parameters = Schema.Struct({
+  path: Schema.optional(
+    Schema.String.annotate({
+      description:
+        "Optional workspace-local path to the finalized plan file. Pass this when you saved the plan somewhere other than the provided plan file path.",
+    }),
+  ),
+})
+
+type Params = Schema.Schema.Type<typeof Parameters>
+
+export const PlanExitTool = Tool.define(
+  "plan_exit",
+  Effect.gen(function* () {
+    const session = yield* Session.Service
+
+    return {
+      description: EXIT_DESCRIPTION,
+      parameters: Parameters,
+      execute: (params: Params, ctx: Tool.Context) =>
+        Effect.gen(function* () {
+          const instance = yield* InstanceState.context
+          const info = yield* session.get(ctx.sessionID)
+          const file = PlanFile.resolve(params.path, instance) ?? Session.plan(info, instance)
+          const plan = PlanFile.display(file, instance)
+          return {
+            title: "Planning complete",
+            output: `Plan is ready at ${plan}. Ending planning turn.`,
+            metadata: { plan },
+          }
+        }).pipe(Effect.orDie),
+    }
+  }),
+)
